Removed STM32 based modulation and moved task to PH of Si4464 (so stable HSE is not needed for STM32 anymore which was needed for radio modulation) Fixed receiver value in decoder.py Limited data display on decoder map to 14 days Implemented additional exception handlers for bad internet connections Modified SSDV/APRS protocol: removed redundant transmitted data Modified SSDV/APRS protocol: Increased amount of data being sent in one packet Adjusted clocks: (in order to save energy) - Use HSI for STM32 only - Use HSE for Si4464 only (HSE switched off, only switched on when Si4464 needs it) - Reduced STM32 clock to 3MHz Improved Morse implementation Fixed thread (module) startup delay bug Changed behavior of Watchdog LED: Flashing instead of blinking (in order to save energy) Fixed preamble bug from AX25 implementation

+++ tracker/software/protocols/aprs/ax25.c
@@ -120,7 +120,7 @@ void ax25_init(ax25_t *packet)
 
 void ax25_send_header(ax25_t *packet, const char *callsign, uint8_t ssid, const char *path, uint16_t preamble)
 {
-	uint8_t i, j;
+	uint16_t i, j;
 	uint8_t tmp[8];
 	packet->ones_in_a_row = 0;
 	packet->crc = 0xffff;
